requirements.txt
pydoll-python
websockets
